split reflection file in validation , serialization & deserialization
