+++ packages/runtype/index.ts
@@ -6,5 +6,8 @@
  * ######## */
 
 export * from './src/constants';
-export * from './src/functionReflection';
+export * from './src/functionValidation';
+export * from './src/functionSerialization';
+export * from './src/functionDeSerialization';
+export * from './src/lib';
 export * from './src/types';
